correct the mapping of Payment Method Type
