+++ crates/common_enums/src/transformers.rs
@@ -2,7 +2,7 @@ use std::fmt::{Display, Formatter};
 
 use serde::{Deserialize, Serialize};
 
-use crate::enums::{Country, CountryAlpha2, CountryAlpha3, PaymentMethod, PaymentMethodType};
+use crate::enums::{Country, CountryAlpha2, CountryAlpha3};
 
 impl Display for NumericCountryCodeParseError {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
@@ -1783,96 +1783,6 @@ impl Country {
             Self::Yemen => 887,
             Self::Zambia => 894,
             Self::Zimbabwe => 716,
-        }
-    }
-}
-
-impl From<PaymentMethodType> for PaymentMethod {
-    fn from(value: PaymentMethodType) -> Self {
-        match value {
-            PaymentMethodType::Ach => Self::BankDebit,
-            PaymentMethodType::Affirm => Self::PayLater,
-            PaymentMethodType::AfterpayClearpay => Self::PayLater,
-            PaymentMethodType::AliPay => Self::Wallet,
-            PaymentMethodType::AliPayHk => Self::Wallet,
-            PaymentMethodType::Alma => Self::PayLater,
-            PaymentMethodType::ApplePay => Self::Wallet,
-            PaymentMethodType::Bacs => Self::BankDebit,
-            PaymentMethodType::BancontactCard => Self::BankRedirect,
-            PaymentMethodType::BcaBankTransfer => Self::BankTransfer,
-            PaymentMethodType::Becs => Self::BankDebit,
-            PaymentMethodType::BniVa => Self::BankTransfer,
-            PaymentMethodType::BriVa => Self::BankTransfer,
-            PaymentMethodType::Benefit => Self::CardRedirect,
-            PaymentMethodType::Bizum => Self::BankRedirect,
-            PaymentMethodType::Blik => Self::BankRedirect,
-            PaymentMethodType::Alfamart => Self::Voucher,
-            PaymentMethodType::CardRedirect => Self::CardRedirect,
-            PaymentMethodType::CimbVa => Self::BankTransfer,
-            PaymentMethodType::ClassicReward => Self::Reward,
-            PaymentMethodType::Credit => Self::Card,
-            PaymentMethodType::CryptoCurrency => Self::Crypto,
-            PaymentMethodType::Dana => Self::Wallet,
-            PaymentMethodType::DanamonVa => Self::BankTransfer,
-            PaymentMethodType::Debit => Self::Card,
-            PaymentMethodType::Eps => Self::BankRedirect,
-            PaymentMethodType::Evoucher => Self::Reward,
-            PaymentMethodType::Giropay => Self::BankRedirect,
-            PaymentMethodType::GooglePay => Self::Wallet,
-            PaymentMethodType::GoPay => Self::Wallet,
-            PaymentMethodType::Gcash => Self::Wallet,
-            PaymentMethodType::Ideal => Self::BankRedirect,
-            PaymentMethodType::Klarna => Self::PayLater,
-            PaymentMethodType::KakaoPay => Self::Wallet,
-            PaymentMethodType::Knet => Self::CardRedirect,
-            PaymentMethodType::MbWay => Self::Wallet,
-            PaymentMethodType::MobilePay => Self::Wallet,
-            PaymentMethodType::Momo => Self::Wallet,
-            PaymentMethodType::MomoAtm => Self::CardRedirect,
-            PaymentMethodType::Multibanco => Self::BankTransfer,
-            PaymentMethodType::MandiriVa => Self::BankTransfer,
-            PaymentMethodType::Interac => Self::BankRedirect,
-            PaymentMethodType::Indomaret => Self::Voucher,
-            PaymentMethodType::OnlineBankingCzechRepublic => Self::BankRedirect,
-            PaymentMethodType::OnlineBankingFinland => Self::BankRedirect,
-            PaymentMethodType::OnlineBankingFpx => Self::BankRedirect,
-            PaymentMethodType::OnlineBankingThailand => Self::BankRedirect,
-            PaymentMethodType::OnlineBankingPoland => Self::BankRedirect,
-            PaymentMethodType::OnlineBankingSlovakia => Self::BankRedirect,
-            PaymentMethodType::PermataBankTransfer => Self::BankTransfer,
-            PaymentMethodType::Pix => Self::BankTransfer,
-            PaymentMethodType::Pse => Self::BankTransfer,
-            PaymentMethodType::PayBright => Self::PayLater,
-            PaymentMethodType::Paypal => Self::Wallet,
-            PaymentMethodType::PaySafeCard => Self::GiftCard,
-            PaymentMethodType::Przelewy24 => Self::BankRedirect,
-            PaymentMethodType::SamsungPay => Self::Wallet,
-            PaymentMethodType::Sepa => Self::BankDebit,
-            PaymentMethodType::Sofort => Self::BankRedirect,
-            PaymentMethodType::Swish => Self::BankRedirect,
-            PaymentMethodType::Trustly => Self::BankRedirect,
-            PaymentMethodType::Twint => Self::Wallet,
-            PaymentMethodType::UpiCollect => Self::Upi,
-            PaymentMethodType::Vipps => Self::Wallet,
-            PaymentMethodType::Walley => Self::PayLater,
-            PaymentMethodType::WeChatPay => Self::Wallet,
-            PaymentMethodType::TouchNGo => Self::Wallet,
-            PaymentMethodType::Atome => Self::PayLater,
-            PaymentMethodType::Boleto => Self::Voucher,
-            PaymentMethodType::Efecty => Self::Voucher,
-            PaymentMethodType::PagoEfectivo => Self::Voucher,
-            PaymentMethodType::RedCompra => Self::Voucher,
-            PaymentMethodType::RedPagos => Self::Voucher,
-            PaymentMethodType::Cashapp => Self::Wallet,
-            PaymentMethodType::Givex => Self::GiftCard,
-            PaymentMethodType::Oxxo => Self::Voucher,
-            PaymentMethodType::OpenBankingUk => Self::BankRedirect,
-            PaymentMethodType::SevenEleven => Self::Voucher,
-            PaymentMethodType::Lawson => Self::Voucher,
-            PaymentMethodType::MiniStop => Self::Voucher,
-            PaymentMethodType::FamilyMart => Self::Voucher,
-            PaymentMethodType::Seicomart => Self::Voucher,
-            PaymentMethodType::PayEasy => Self::Voucher,
         }
     }
 }
