+++ crates/router/src/core/payments/operations/payment_session.rs
@@ -418,7 +418,11 @@ where
                                 is_invoke_sdk_client && is_sent_in_request
                             })
                             .map(|payment_method_type| {
-                                (connector_account, payment_method_type.payment_method_type)
+                                (
+                                    connector_account,
+                                    parsed_payment_methods_enabled.payment_method,
+                                    payment_method_type.payment_method_type,
+                                )
                             })
                             .collect::<Vec<_>>()
                     })
@@ -429,7 +433,7 @@ where
         let mut session_connector_data =
             Vec::with_capacity(connector_and_supporting_payment_method_type.len());
 
-        for (merchant_connector_account, payment_method_type) in
+        for (merchant_connector_account, payment_method, payment_method_type) in
             connector_and_supporting_payment_method_type
         {
             let connector_type = api::GetToken::from(payment_method_type);
@@ -444,6 +448,7 @@ where
                 err
             }) {
                 session_connector_data.push(api::SessionConnectorData {
+                    payment_method,
                     payment_method_type,
                     connector: connector_data,
                     business_sub_label: merchant_connector_account.business_sub_label.clone(),
