+++ crates/router/src/types/api/payment_methods.rs
@@ -1,4 +1,4 @@
-use api_models::enums as api_enums;
+// use api_models::enums as api_enums;
 pub use api_models::payment_methods::{
     CardDetail, CardDetailFromLocker, CardDetailsPaymentMethod, CustomerPaymentMethod,
     CustomerPaymentMethodsListResponse, DeleteTokenizeByTokenRequest, GetTokenizePayloadRequest,
@@ -9,9 +9,9 @@ pub use api_models::payment_methods::{
 };
 use error_stack::report;
 
-use crate::{
-    core::errors::{self, RouterResult},
-    types::transformers::ForeignFrom,
+use crate::core::{
+    errors::{self, RouterResult},
+    utils,
 };
 
 pub(crate) trait PaymentMethodCreateExt {
@@ -21,16 +21,18 @@ pub(crate) trait PaymentMethodCreateExt {
 // convert self.payment_method_type to payment_method and compare it against self.payment_method
 impl PaymentMethodCreateExt for PaymentMethodCreate {
     fn validate(&self) -> RouterResult<()> {
-        let payment_method: Option<api_enums::PaymentMethod> =
-            self.payment_method_type.map(ForeignFrom::foreign_from);
-        if payment_method
-            .map(|payment_method| payment_method != self.payment_method)
-            .unwrap_or(false)
-        {
-            return Err(report!(errors::ApiErrorResponse::InvalidRequestData {
-                message: "Invalid 'payment_method_type' provided".to_string()
-            })
-            .attach_printable("Invalid payment method type"));
+        // let payment_method: Option<api_enums::PaymentMethod> =
+        //     self.payment_method_type.map(ForeignFrom::foreign_from);
+        if let Some(payment_method_type) = self.payment_method_type {
+            if !utils::validate_payment_method_type_against_payment_method(
+                self.payment_method,
+                payment_method_type,
+            ) {
+                return Err(report!(errors::ApiErrorResponse::InvalidRequestData {
+                    message: "Invalid 'payment_method_type' provided".to_string()
+                })
+                .attach_printable("Invalid payment method type"));
+            }
         }
         Ok(())
     }
