+++ crates/router/src/types/transformers.rs
@@ -361,98 +361,6 @@ impl ForeignFrom<api_enums::IntentStatus> for Option<storage_enums::EventType> {
     }
 }
 
-impl ForeignFrom<api_enums::PaymentMethodType> for api_enums::PaymentMethod {
-    fn foreign_from(payment_method_type: api_enums::PaymentMethodType) -> Self {
-        match payment_method_type {
-            api_enums::PaymentMethodType::ApplePay
-            | api_enums::PaymentMethodType::GooglePay
-            | api_enums::PaymentMethodType::Paypal
-            | api_enums::PaymentMethodType::AliPay
-            | api_enums::PaymentMethodType::AliPayHk
-            | api_enums::PaymentMethodType::Dana
-            | api_enums::PaymentMethodType::MbWay
-            | api_enums::PaymentMethodType::MobilePay
-            | api_enums::PaymentMethodType::SamsungPay
-            | api_enums::PaymentMethodType::Twint
-            | api_enums::PaymentMethodType::Vipps
-            | api_enums::PaymentMethodType::TouchNGo
-            | api_enums::PaymentMethodType::Swish
-            | api_enums::PaymentMethodType::WeChatPay
-            | api_enums::PaymentMethodType::GoPay
-            | api_enums::PaymentMethodType::Gcash
-            | api_enums::PaymentMethodType::Momo
-            | api_enums::PaymentMethodType::Cashapp
-            | api_enums::PaymentMethodType::KakaoPay => Self::Wallet,
-            api_enums::PaymentMethodType::Affirm
-            | api_enums::PaymentMethodType::Alma
-            | api_enums::PaymentMethodType::AfterpayClearpay
-            | api_enums::PaymentMethodType::Klarna
-            | api_enums::PaymentMethodType::PayBright
-            | api_enums::PaymentMethodType::Atome
-            | api_enums::PaymentMethodType::Walley => Self::PayLater,
-            api_enums::PaymentMethodType::Giropay
-            | api_enums::PaymentMethodType::Ideal
-            | api_enums::PaymentMethodType::Sofort
-            | api_enums::PaymentMethodType::Eps
-            | api_enums::PaymentMethodType::BancontactCard
-            | api_enums::PaymentMethodType::Blik
-            | api_enums::PaymentMethodType::OnlineBankingThailand
-            | api_enums::PaymentMethodType::OnlineBankingCzechRepublic
-            | api_enums::PaymentMethodType::OnlineBankingFinland
-            | api_enums::PaymentMethodType::OnlineBankingFpx
-            | api_enums::PaymentMethodType::OnlineBankingPoland
-            | api_enums::PaymentMethodType::OnlineBankingSlovakia
-            | api_enums::PaymentMethodType::OpenBankingUk
-            | api_enums::PaymentMethodType::Przelewy24
-            | api_enums::PaymentMethodType::Trustly
-            | api_enums::PaymentMethodType::Bizum
-            | api_enums::PaymentMethodType::Interac => Self::BankRedirect,
-            api_enums::PaymentMethodType::UpiCollect => Self::Upi,
-            api_enums::PaymentMethodType::CryptoCurrency => Self::Crypto,
-            api_enums::PaymentMethodType::Ach
-            | api_enums::PaymentMethodType::Sepa
-            | api_enums::PaymentMethodType::Bacs
-            | api_enums::PaymentMethodType::Becs => Self::BankDebit,
-            api_enums::PaymentMethodType::Credit | api_enums::PaymentMethodType::Debit => {
-                Self::Card
-            }
-            api_enums::PaymentMethodType::Evoucher
-            | api_enums::PaymentMethodType::ClassicReward => Self::Reward,
-            api_enums::PaymentMethodType::Boleto
-            | api_enums::PaymentMethodType::Efecty
-            | api_enums::PaymentMethodType::PagoEfectivo
-            | api_enums::PaymentMethodType::RedCompra
-            | api_enums::PaymentMethodType::Alfamart
-            | api_enums::PaymentMethodType::Indomaret
-            | api_enums::PaymentMethodType::Oxxo
-            | api_enums::PaymentMethodType::SevenEleven
-            | api_enums::PaymentMethodType::Lawson
-            | api_enums::PaymentMethodType::MiniStop
-            | api_enums::PaymentMethodType::FamilyMart
-            | api_enums::PaymentMethodType::Seicomart
-            | api_enums::PaymentMethodType::PayEasy
-            | api_enums::PaymentMethodType::RedPagos => Self::Voucher,
-            api_enums::PaymentMethodType::Pse
-            | api_enums::PaymentMethodType::Multibanco
-            | api_enums::PaymentMethodType::PermataBankTransfer
-            | api_enums::PaymentMethodType::BcaBankTransfer
-            | api_enums::PaymentMethodType::BniVa
-            | api_enums::PaymentMethodType::BriVa
-            | api_enums::PaymentMethodType::CimbVa
-            | api_enums::PaymentMethodType::DanamonVa
-            | api_enums::PaymentMethodType::MandiriVa
-            | api_enums::PaymentMethodType::Pix => Self::BankTransfer,
-            api_enums::PaymentMethodType::Givex | api_enums::PaymentMethodType::PaySafeCard => {
-                Self::GiftCard
-            }
-            api_enums::PaymentMethodType::Benefit
-            | api_enums::PaymentMethodType::Knet
-            | api_enums::PaymentMethodType::MomoAtm
-            | api_enums::PaymentMethodType::CardRedirect => Self::CardRedirect,
-        }
-    }
-}
-
 impl ForeignTryFrom<api_models::payments::PaymentMethodData> for api_enums::PaymentMethod {
     type Error = errors::ApiErrorResponse;
     fn foreign_try_from(
